+++ src/domains/MyPage/pages/ProfilePage.tsx
@@ -2,20 +2,24 @@ import BadgeModal from '@/domains/MyPage/components/profile/BadgeModal';
 import UserProfile from '@/domains/MyPage/components/profile/UserProfile';
 import { BADGES, USAGE_HISTORY } from '@/domains/MyPage/constants/profile';
 import type { UserInfo } from '@/domains/MyPage/types/profile';
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import outstandingIcon from '@/assets/icons/outstanding_icon.png';
 import UsageHistory from '@/domains/MyPage/components/profile/UsageHistory';
-
-const Breadcrumb: React.FC = () => (
-  <div className="flex">
-    <p className="text-gray-400">마이페이지</p>&nbsp;/&nbsp;<p>내 정보</p>
-  </div>
-);
+import { getUserInfo } from '@/domains/MyPage/api/profile';
+import { Breadcrumb } from '@/domains/MyPage/components/Breadcrumb';
 
 const ProfilePage: React.FC = () => {
   const [open, setOpen] = useState<boolean>(false);
   const [selectedBadge, setSelectedBadge] = useState<string>('earlybird');
   const [tempBadge, setTempBadge] = useState<string>(selectedBadge);
+
+  useEffect(() => {
+    async function fetchData() {
+      const data = await getUserInfo();
+      console.log(data);
+    }
+    fetchData();
+  }, []);
 
   // 실제로는 API에서 받아올 데이터
   const userInfo: UserInfo = {
@@ -51,7 +55,7 @@ const ProfilePage: React.FC = () => {
   return (
     <>
       <div className="w-full max-w-[1050px]">
-        <Breadcrumb />
+        <Breadcrumb title="내 정보" />
 
         <div>
           <div className="text-[32px]">내 정보</div>
